Add cluster model

+++ cephlcm/common/models/server.py
@@ -61,16 +61,52 @@ class ServerModel(generic.Model):
         model.ip = ip
         model.facts = facts or {}
         model.cluster_id = cluster_id
+        model._cluster = None
         model.state = state
         model.initiator_id = initiator_id
         model.save()
 
         return model
 
+    @classmethod
+    def cluster_servers(cls, cluster_id):
+        query = {
+            "cluster_id": cluster_id,
+            "is_latest": True,
+            "time_deleted": 0
+        }
+
+        return cls.list_raw(query)
+
     @property
     def cluster(self):
-        # TODO(Sergey Arkhipov): Implement after Cluster model will be add
-        return {}
+        if self._cluster:
+            return self._cluster
+
+        from cephlcm.common.models import cluster
+
+        self._cluster = cluster.ClusterModel.find_by_model_id(self.cluster_id)
+
+        return self._cluster
+
+    @cluster.setter
+    def cluster(self, value):
+        new_cluster_id = None
+
+        if hasattr(value, "model_id"):
+            new_cluster_id = value.model_id
+        elif isinstance(value, dict):
+            new_cluster_id = value["id"]
+        else:
+            new_cluster_id = value
+
+        if self.cluster_id is not None and new_cluster_id is not None:
+            raise ValueError(
+                "Already defined cluster {0}. "
+                "Set to None first".format(self.cluster_id))
+
+        self.cluster_id = new_cluster_id
+        self._cluster = None
 
     @property
     def state(self):
@@ -117,7 +153,10 @@ class ServerModel(generic.Model):
         self._cluster = None
 
     def delete(self):
-        # TODO(Sergey Arkhipov): After create of cluster model, implement.
+        if self.cluster_id:
+            # TODO(Sergey Arkhipov): Raise proper exception
+            raise Exception
+
         super().delete()
 
     def make_db_document_specific_fields(self):
